requirements.txt
semantic-kernel>=0.9.1
openai>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
